// nodes/CapmonsterCloud/types.ts
export type TaskType =
	| 'json'
	| 'recaptchaV2Enterprise'
	| 'recaptchaV3Enterprise'
	| 'recaptchaV2'
	| 'recaptchaV3'
	| 'imageToText'
	| 'geetest'
	| 'recaptchaClick'
	| 'turnstile'
	| 'turnstileChallenge'
	| 'turnstileWaitRoom'
	| 'complexImageRecognition'
	| 'datadome'
	| 'basilisk'
	| 'tendi'
	| 'amazon'
	| 'binance'
	| 'imperva'
	| 'prosopo'
	| 'temu'
	| 'yidun'
	| 'mtcaptcha'
	| 'altcha'
	| 'funcaptcha'
	| 'castle'
	| 'tspd'
	| 'hunt';